'Affects' command enhancements:
- move 'affects' command into a separate file
- display mana/heal bonus and anti-bonus
- display bonuses to skill learned %%
- display bonuses to skill levels

Also:
- add helper methods to GlobalArray and GlobalRegistryElement
- Character::getConfig() no longer allocates a new PlayerConfig instance
  each time it's called (wtf)
