+++ plug-ins/skills_impl/defaultskillgroup.cpp
@@ -127,7 +127,7 @@ void DefaultSkillGroup::listSkills( PCharacter *ch, ostringstream &buf ) const
 {
     // True if it's a help json dump and not a player requesting the article.
     bool autodump = ch->desc == 0;
-    bool fRus = ch->getConfig()->ruskills;
+    bool fRus = ch->getConfig().ruskills;
     int columns = 3;
     const char *pattern = fRus ? "{%c%-26s{x" : "{%c%-20s{x";
     const char *autopattern = "%-26s";
