+++ src/gref/globalarray.h
@@ -9,6 +9,7 @@
 #include "globalregistry.h"
 
 class GlobalBitvector;
+class StringList;
 
 class GlobalArray : public vector<int> {
 public:
@@ -19,8 +20,12 @@ public:
     virtual ~GlobalArray( );
     
     void clear( );
+    bool isEmpty() const;
     int & operator [] (size_type);
     void applyBitvector(const GlobalBitvector &bitvector, int modifier);
+
+    /** Generate a list of strings describing how much each element is affected: "xxx by 10, yyy by -5". */
+    StringList toStringList(bool fRussian) const;
 
 protected:
     GlobalRegistryBase *registry;
